Student attempt of the B2B project from 2017.
FromBiH is an imaginary platform for the domestic market and companies, in order to strengthen their competitiveness and productivity.

Currently uploaded with the goal of filling the portfolio.

Made in HTML, CSS, PHP and JS (jQuery) with Bootstrap.

The project was presented to the employees of BBI Bank with the aim of finding sponsors to continue it.